IncreasingUsersPerSecProfile#startingFrom should take a Double, not an Int, close #3645
Motivation:

startingFrom is internally a Double but DSL method exposes an Int.

Modification:

Fix parameter type

Result:

startingFrom now takes a Double as expected

+++ gatling-core/src/main/scala/io/gatling/core/controller/inject/open/OpenInjectionSupport.scala
@@ -66,7 +66,7 @@ trait MetaOpenInjectionSupport {
       startingUsers: Double,
       rampDuration:  FiniteDuration
   ) extends MetaInjectionProfile {
-    def startingFrom(startingUsers: Int): IncreasingUsersPerSecProfile = this.copy(startingUsers = startingUsers)
+    def startingFrom(startingUsers: Double): IncreasingUsersPerSecProfile = this.copy(startingUsers = startingUsers)
     def separatedByRampsLasting(duration: FiniteDuration): IncreasingUsersPerSecProfile = this.copy(rampDuration = duration)
 
     private[inject] def getInjectionSteps: Iterable[OpenInjectionStep] =
